+++ src/pages/Skills.tsx
@@ -1,0 +1,12 @@
+// src/components/AboutMy.tsx
+export default function Skills() {
+    return (
+      <div className="flex items-center justify-center h-screen">
+        <div className="text-center">
+          <h1 className="text-black text-4xl font-bold mb-4">Skills</h1>
+          <p className="text-black text-lg">Información sobre mís erramientas tecnologicas...</p>
+        </div>
+      </div>
+    );
+  }
+
